Bug 572605 - Improve performance for name space wiring lookup
NamespaceList is introduced to index a flat list of elements which have
name spaces.  The NamespaceList assumes the element list is ordered such
that all elements with the same name space are order in one continuous
sequence.  The indexes list contains only the beginning index for each
names pace. This allows a much smaller list to be traversed in order to
find the beginning index of a name space for lookup.  The ending index
is found by looking at the next index.  If the start index is the last
index in the indexes list then the end index for that namespace is the
size of the elements list.

NamespaceList is immutable.  Any updates require a copy of the full list
to be made and modified and then a new NamespaceList to be created.

Additional fixes are included to properly update the requirements and
capabilities for a host wiring when fragments are dynamically attached.

Change-Id: Iff2856fca7aaf5ab9fd1efe08b8778408860a146
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>